# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DMS - Document Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assignedToMe
          in: query
          schema:
            type: ["boolean", "null"]
        - name: protocolId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dossierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: priority
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List workflow definitions
      responses:
        "200":
          description: Workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get workflow definition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /document-workflows:
    get:
      operationId: listDocumentWorkflows
      tags: [workflows]
      summary: List active document workflow instances
      parameters:
        - name: documentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Active workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentWorkflow"
    post:
      operationId: startDocumentWorkflow
      tags: [workflows]
      summary: Start a workflow on a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartWorkflowInput"
      responses:
        "201":
          description: Started workflow instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentWorkflow"
  /document-workflows/{id}/advance:
    post:
      operationId: advanceWorkflowStep
      tags: [workflows]
      summary: Advance or reject a workflow step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowStepOutcome"
      responses:
        "200":
          description: Updated workflow instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentWorkflow"

  # --- SIGNATURES ---
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    TaskList:
      type: object
      required:
        - items
        - total
        - page
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Task:
      type: object
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        progress:
          type: integer
          default: 0
        protocolId:
          type: ["integer", "null"]
        protocolNumber:
          type: ["string", "null"]
        documentId:
          type: ["integer", "null"]
        documentTitle:
          type: ["string", "null"]
        dossierId:
          type: ["integer", "null"]
        dossierTitle:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: string
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required:
        - title
        - priority
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
        protocolId:
          type: ["integer", "null"]
        documentId:
          type: ["integer", "null"]
        dossierId:
          type: ["integer", "null"]
        assignedToId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        progress:
          type: integer
        assignedToId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: string
        outcome:
          type: string

    # --- WORKFLOWS ---
    Workflow:
      type: object
      required:
        - id
        - name
        - steps
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        documentType:
          type: ["string", "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
    WorkflowStep:
      type: object
      required:
        - id
        - name
        - order
        - action
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        action:
          type: string
        responsibleRole:
          type: ["string", "null"]
        dueDays:
          type: ["integer", "null"]
        requiredApprovalType:
          type: string
          default: single
    WorkflowInput:
      type: object
      required:
        - name
        - steps
      properties:
        name:
          type: string
        description:
          type: string
        documentType:
          type: string
        steps:
          type: array
          items:
            type: object
            required:
              - name
              - order
              - action
            properties:
              name:
                type: string
              description:
                type: string
              order:
                type: integer
              action:
                type: string
              responsibleRole:
                type: string
              dueDays:
                type: integer
    DocumentWorkflow:
      type: object
      required:
        - id
        - documentId
        - workflowId
        - workflowName
        - currentStep
        - status
        - startedAt
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentTitle:
          type: ["string", "null"]
        workflowId:
          type: integer
        workflowName:
          type: string
        currentStep:
          type: integer
        totalSteps:
          type: integer
        status:
          type: string
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        currentStepName:
          type: ["string", "null"]
        currentStepAction:
          type: ["string", "null"]
        currentAssigneeId:
          type: ["integer", "null"]
        currentAssigneeName:
          type: ["string", "null"]
    StartWorkflowInput:
      type: object
      required:
        - documentId
        - workflowId
      properties:
        documentId:
          type: integer
        workflowId:
          type: integer
        note:
          type: string
    WorkflowStepOutcome:
      type: object
      required:
        - outcome
      properties:
        outcome:
          type: string
        note:
          type: string
        delegateToId:
          type: ["integer", "null"]

    # --- SIGNATURES ---
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
        - isActive
        - createdAt
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        area:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required:
        - email
        - name
        - role
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
        area:
          type: string
        section:
          type: string
        avatarUrl:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        area:
          type: string
        section:
          type: string
        isActive:
          type: boolean
        avatarUrl:
          type: string

    # --- CLASSIFICATIONS ---
...