# lib/api-zod/src/generated/types/workflowStepOutcome.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMS - Document Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkflowStepOutcome {
  outcome: string;
  note?: string;
  /** @nullable */
  delegateToId?: number | null;
}